navigation

+++ app/src/main/java/com/joelkanyi/composenavigation_assignment/Screen/Rating_Screen.kt
@@ -21,7 +21,7 @@ import androidx.navigation.NavController
 @Composable
 fun Rating_Screen(
 
-    id:Int,
+    id:Float,
     navController: NavController
 ) {
     Scaffold(
